refactor slightly
