+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/lib/AprilTagDetectionPipeline.java
@@ -19,7 +19,7 @@
  * SOFTWARE.
  */
 
-package org.firstinspires.ftc.teamcode.apriltagvision;
+package org.firstinspires.ftc.teamcode.lib;
 
 import org.opencv.calib3d.Calib3d;
 import org.opencv.core.CvType;
